requirements.txt
contextily
geopandas
google-cloud
holidays
ipykernel
matplotlib
seaborn
plotly
nbformat